requirements.txt
absl-py
altair
annotated-types
anyio
astunparse
attrs
beautifulsoup4
blinker
cachetools
certifi
charset-normalizer
click
colorama
exceptiongroup
fastapi
filelock
flatbuffers
gast
gdown
python-multipart
gitdb
GitPython
google-auth
google-auth-oauthlib
google-pasta
grpcio
h11
h5py
idna
importlib_metadata
Jinja2
joblib
jsonschema
jsonschema-specifications
keras
libclang
Markdown
markdown-it-py
MarkupSafe
mdurl
ml-dtypes
narwhals
numpy
oauthlib
opt_einsum
packaging
pandas
pillow
protobuf
pyarrow
pyasn1
pyasn1_modules
pydantic
pydantic_core
pydeck
Pygments
PySocks
python-dateutil
python-multipart
pytz
referencing
requests
requests-oauthlib
rich
rpds-py
rsa
scikit-learn
scipy
six
smmap
sniffio
soupsieve
starlette
streamlit
tenacity
tensorboard
tensorboard-data-server
tensorflow
tensorflow-estimator
tensorflow-intel
tensorflow-io-gcs-filesystem
termcolor
threadpoolctl
toml
tornado
tqdm
typing_extensions
tzdata
tzlocal
urllib3
uvicorn
validators
watchdog
Werkzeug
wrapt
zipp

